+++ docs/changelog.md
@@ -8,9 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
-## 2025-12-11
+## 2026-01-07
 
 * Change the naming scheme for the bacula-fd charm
+
+## 2025-12-17
+
+* Moved charm architecture documentation from Explanation to Reference category.
 
 ## 2025-12-09